+++ users/models.py
@@ -1,4 +1,7 @@
-from datetime import datetime
+from datetime import (
+    datetime,
+    timedelta,
+)
 
 from django.contrib.auth.models import (
     AbstractUser,
@@ -6,6 +9,7 @@ from django.contrib.auth.models import (
     UserManager,
 )
 from django.db import models
+from django.utils import timezone
 
 
 def current_year() -> int:
@@ -24,6 +28,15 @@ class User(AbstractUser):
 
     groups: 'models.manager.RelatedManager[Group]'
 
+    is_verified: 'models.BooleanField[bool, bool]' = models.BooleanField(
+        'e-mail verificado', default=False,
+    )
+
+    verify_email_sent = models.DateTimeField(
+        'último envío de correo de verificación',
+        blank=True, null=True,
+    )
+
     telegram_user: 'models.CharField[str, str]' = models.CharField(
         'usuario de telegram', max_length=64, blank=True
     )
@@ -37,3 +50,11 @@ class User(AbstractUser):
 
     def __str__(self):
         return f'{self.first_name} {self.last_name} - {self.email}'
+
+    def can_send_verify_email(self) -> bool:
+        """Checks if the user can send a verify email.
+        """
+        return (
+            self.verify_email_sent is None
+            or self.verify_email_sent + timedelta(seconds=60) < timezone.now()
+        )